Moved some code from main to distro
