--- distro/sender.go
@@ -1,5 +1,0 @@
-package distro
-
-type Sender interface {
-	Send(data string)
-}
